Refactor onboarding
