--- app/components/UI/Rewards/components/Onboarding/OnboardingStep2.tsx
@@ -1,78 +1,0 @@
-import React, { useCallback } from 'react';
-import { Image, useWindowDimensions } from 'react-native';
-import { useNavigation } from '@react-navigation/native';
-import { useDispatch } from 'react-redux';
-import { useTailwind } from '@metamask/design-system-twrnc-preset';
-import Routes from '../../../../../constants/navigation/Routes';
-import { Box, Text, TextVariant } from '@metamask/design-system-react-native';
-import step2Img from '../../../../../images/rewards/rewards-onboarding-step2.png';
-import Step2BgImg from '../../../../../images/rewards/rewards-onboarding-step2-bg.svg';
-import { setOnboardingActiveStep } from '../../../../../actions/rewards';
-import { OnboardingStep } from '../../../../../reducers/rewards/types';
-import { useTheme } from '../../../../../util/theme';
-import { strings } from '../../../../../../locales/i18n';
-import OnboardingStepComponent from './OnboardingStep';
-
-const OnboardingStep2: React.FC = () => {
-  const navigation = useNavigation();
-  const dispatch = useDispatch();
-  const tw = useTailwind();
-  const { colors } = useTheme();
-  const { width: screenWidth, height: screenHeight } = useWindowDimensions();
-
-  const handleNext = useCallback(() => {
-    dispatch(setOnboardingActiveStep(OnboardingStep.STEP_3));
-    navigation.navigate(Routes.REWARDS_ONBOARDING_3);
-  }, [dispatch, navigation]);
-
-  const handleSkip = useCallback(() => {
-    dispatch(setOnboardingActiveStep(OnboardingStep.STEP_4));
-    navigation.navigate(Routes.REWARDS_ONBOARDING_4);
-  }, [dispatch, navigation]);
-
-  const renderStepImage = () => (
-    <>
-      <Step2BgImg
-        name="rewards-onboarding-step2-bg"
-        fill={colors.background.muted}
-        style={tw.style('absolute flex-1')}
-        width={screenWidth}
-        height={screenHeight}
-      />
-
-      <Image
-        source={step2Img}
-        style={tw.style('h-[75%] z-10')}
-        testID="step-2-image"
-        resizeMode="contain"
-      />
-    </>
-  );
-
-  const renderStepInfo = () => (
-    <Box twClassName="flex-col gap-2 min-h-30">
-      <Text variant={TextVariant.HeadingLg} twClassName="text-center">
-        {strings('rewards.onboarding.step2_title')}
-      </Text>
-
-      <Text
-        variant={TextVariant.BodyMd}
-        twClassName="text-center text-alternative"
-      >
-        {strings('rewards.onboarding.step2_description')}
-      </Text>
-    </Box>
-  );
-
-  return (
-    <OnboardingStepComponent
-      currentStep={2}
-      onNext={handleNext}
-      onSkip={handleSkip}
-      renderStepImage={renderStepImage}
-      renderStepInfo={renderStepInfo}
-    />
-  );
-};
-
-export default OnboardingStep2;
