+++ app/components/UI/Rewards/components/Onboarding/__tests__/OnboardingStep.test.tsx
@@ -1,11 +1,7 @@
 import React from 'react';
-import { screen, fireEvent, waitFor } from '@testing-library/react-native';
-import { Text, Linking, PanResponder } from 'react-native';
+import { screen, fireEvent } from '@testing-library/react-native';
+import { Text, PanResponder } from 'react-native';
 import OnboardingStep from '../OnboardingStep';
-import OnboardingStep1 from '../OnboardingStep1';
-import OnboardingStep2 from '../OnboardingStep2';
-import OnboardingStep3 from '../OnboardingStep3';
-import OnboardingStep4 from '../OnboardingStep4';
 import { renderWithProviders } from '../testUtils';
 import { REWARDS_VIEW_SELECTORS } from '../../../Views/RewardsView.constants';
 
@@ -418,7 +414,7 @@ describe('image and background rendering', () => {
   });
 });
 
-describe('nextButtonAlternative rendering', () => {
+describe('renderLegalDisclaimer rendering', () => {
   it('should render alternative button when provided', () => {
     const mockAlternativeButton = jest.fn(() => (
       <Text testID="alternative-button">Alternative Action</Text>
@@ -426,7 +422,7 @@ describe('nextButtonAlternative rendering', () => {
 
     renderWithProviders(
       <OnboardingStep
-        nextButtonAlternative={mockAlternativeButton}
+        renderLegalDisclaimer={mockAlternativeButton}
         {...defaultProps}
       />,
     );
@@ -754,7 +750,7 @@ describe('component composition and layout', () => {
         {...defaultProps}
         renderStepImage={mockRenderStepImage}
         renderStepInfo={mockRenderStepInfoComp}
-        nextButtonAlternative={mockAlternativeButton}
+        renderLegalDisclaimer={mockAlternativeButton}
         nextButtonText="Custom Next"
       />,
     );
@@ -765,461 +761,5 @@ describe('component composition and layout', () => {
     expect(screen.getByTestId('step-info')).toBeDefined();
     expect(screen.getByText('Custom Next')).toBeDefined();
     expect(screen.getByTestId('alternative-button')).toBeDefined();
-  });
-});
-
-// Individual OnboardingStep Component Tests
-describe('OnboardingStep1', () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-  });
-
-  describe('rendering', () => {
-    it('should render without crashing', () => {
-      renderWithProviders(<OnboardingStep1 />);
-      expect(screen.getByTestId('onboarding-step-container')).toBeDefined();
-    });
-
-    it('should render step 1 image with correct testID', () => {
-      renderWithProviders(<OnboardingStep1 />);
-      expect(screen.getByTestId('step-1-image')).toBeDefined();
-    });
-
-    it('should display step 1 title and description', () => {
-      renderWithProviders(<OnboardingStep1 />);
-      expect(
-        screen.getByText('mocked_rewards.onboarding.step1_title'),
-      ).toBeDefined();
-      expect(
-        screen.getByText('mocked_rewards.onboarding.step1_description'),
-      ).toBeDefined();
-    });
-
-    it('should show progress indicator with currentStep set to 1', () => {
-      renderWithProviders(<OnboardingStep1 />);
-      expect(screen.getByTestId('progress-indicator-container')).toBeDefined();
-    });
-
-    it('should render skip and next buttons', () => {
-      renderWithProviders(<OnboardingStep1 />);
-      expect(
-        screen.getByText('mocked_rewards.onboarding.step_confirm'),
-      ).toBeDefined();
-    });
-  });
-
-  describe('navigation', () => {
-    it('should navigate to step 2 when next button is pressed', () => {
-      renderWithProviders(<OnboardingStep1 />);
-
-      const nextButton = screen.getByText(
-        'mocked_rewards.onboarding.step_confirm',
-      );
-      fireEvent.press(nextButton);
-
-      expect(mockDispatch).toHaveBeenCalledWith(
-        expect.objectContaining({
-          type: expect.any(String),
-        }),
-      );
-      expect(mockNavigate).toHaveBeenCalledWith('RewardsOnboarding2');
-    });
-  });
-
-  describe('content verification', () => {
-    it('should use correct images and background elements', () => {
-      renderWithProviders(<OnboardingStep1 />);
-
-      // Verify step image is present
-      const stepImage = screen.getByTestId('step-1-image');
-      expect(stepImage).toBeDefined();
-      expect(stepImage.props.resizeMode).toBe('contain');
-    });
-  });
-});
-
-describe('OnboardingStep2', () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-  });
-
-  describe('rendering', () => {
-    it('should render without crashing', () => {
-      renderWithProviders(<OnboardingStep2 />);
-      expect(screen.getByTestId('onboarding-step-container')).toBeDefined();
-    });
-
-    it('should render step 2 image with correct testID', () => {
-      renderWithProviders(<OnboardingStep2 />);
-      expect(screen.getByTestId('step-2-image')).toBeDefined();
-    });
-
-    it('should display step 2 title and description', () => {
-      renderWithProviders(<OnboardingStep2 />);
-      expect(
-        screen.getByText('mocked_rewards.onboarding.step2_title'),
-      ).toBeDefined();
-      expect(
-        screen.getByText('mocked_rewards.onboarding.step2_description'),
-      ).toBeDefined();
-    });
-
-    it('should show progress indicator with currentStep set to 2', () => {
-      renderWithProviders(<OnboardingStep2 />);
-      expect(screen.getByTestId('progress-indicator-container')).toBeDefined();
-    });
-  });
-
-  describe('navigation', () => {
-    it('should navigate to step 3 when next button is pressed', () => {
-      renderWithProviders(<OnboardingStep2 />);
-
-      const nextButton = screen.getByText(
-        'mocked_rewards.onboarding.step_confirm',
-      );
-      fireEvent.press(nextButton);
-
-      expect(mockDispatch).toHaveBeenCalledWith(
-        expect.objectContaining({
-          type: expect.any(String),
-        }),
-      );
-      expect(mockNavigate).toHaveBeenCalledWith('RewardsOnboarding3');
-    });
-  });
-
-  describe('content verification', () => {
-    it('should use correct images and background elements', () => {
-      renderWithProviders(<OnboardingStep2 />);
-
-      const stepImage = screen.getByTestId('step-2-image');
-      expect(stepImage).toBeDefined();
-      expect(stepImage.props.resizeMode).toBe('contain');
-    });
-  });
-});
-
-describe('OnboardingStep3', () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-  });
-
-  describe('rendering', () => {
-    it('should render without crashing', () => {
-      renderWithProviders(<OnboardingStep3 />);
-      expect(screen.getByTestId('onboarding-step-container')).toBeDefined();
-    });
-
-    it('should render step 3 image with correct testID', () => {
-      renderWithProviders(<OnboardingStep3 />);
-      expect(screen.getByTestId('step-3-image')).toBeDefined();
-    });
-
-    it('should display step 3 title and description', () => {
-      renderWithProviders(<OnboardingStep3 />);
-      expect(
-        screen.getByText('mocked_rewards.onboarding.step3_title'),
-      ).toBeDefined();
-      expect(
-        screen.getByText('mocked_rewards.onboarding.step3_description'),
-      ).toBeDefined();
-    });
-
-    it('should show progress indicator with currentStep set to 3', () => {
-      renderWithProviders(<OnboardingStep3 />);
-      expect(screen.getByTestId('progress-indicator-container')).toBeDefined();
-    });
-  });
-
-  describe('navigation', () => {
-    it('should navigate to step 4 when next button is pressed', async () => {
-      renderWithProviders(<OnboardingStep3 />);
-
-      const nextButton = screen.getByText(
-        'mocked_rewards.onboarding.step_confirm',
-      );
-      fireEvent.press(nextButton);
-
-      await waitFor(() => {
-        expect(mockDispatch).toHaveBeenCalledWith(
-          expect.objectContaining({
-            type: expect.any(String),
-          }),
-        );
-        expect(mockNavigate).toHaveBeenCalledWith('RewardsOnboarding4');
-      });
-    });
-  });
-
-  describe('content verification', () => {
-    it('should use correct images and background elements', () => {
-      renderWithProviders(<OnboardingStep3 />);
-
-      const stepImage = screen.getByTestId('step-3-image');
-      expect(stepImage).toBeDefined();
-      expect(stepImage.props.resizeMode).toBe('contain');
-    });
-  });
-});
-
-describe('OnboardingStep4', () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-    // Reset mock state
-    mockUseOptin.optinError = null;
-    mockUseOptin.optinLoading = false;
-    mockUseValidateReferralCode.referralCode = '';
-    mockUseValidateReferralCode.isValidating = false;
-    mockUseValidateReferralCode.isValid = false;
-  });
-
-  describe('rendering', () => {
-    it('should render without crashing', () => {
-      renderWithProviders(<OnboardingStep4 />);
-      expect(screen.getByTestId('onboarding-step-container')).toBeDefined();
-    });
-
-    it('should render step 4 image with correct testID', () => {
-      renderWithProviders(<OnboardingStep4 />);
-      expect(screen.getByTestId('step-4-image')).toBeDefined();
-    });
-
-    it('should display default title when no referral code is valid', () => {
-      renderWithProviders(<OnboardingStep4 />);
-      expect(
-        screen.getByText('mocked_rewards.onboarding.step4_title'),
-      ).toBeDefined();
-    });
-
-    it('should display referral bonus title when referral code is valid', () => {
-      mockUseValidateReferralCode.isValid = true;
-
-      renderWithProviders(<OnboardingStep4 />);
-      expect(
-        screen.getByText(
-          'mocked_rewards.onboarding.step4_title_referral_bonus',
-        ),
-      ).toBeDefined();
-    });
-
-    it('should render referral code input field', () => {
-      renderWithProviders(<OnboardingStep4 />);
-      expect(screen.getByDisplayValue('')).toBeDefined(); // TextField should be present
-    });
-
-    it('should render legal disclaimer with learn more link', () => {
-      renderWithProviders(<OnboardingStep4 />);
-
-      // The component renders a Text component with nested Text components
-      // Instead of looking for exact mock keys, we'll check for the presence of Text elements
-      const textElements = screen.getAllByText(
-        /mocked_rewards.onboarding.step4_legal_disclaimer/,
-      );
-
-      // Verify we have at least one text element for the legal disclaimer
-      expect(textElements.length).toBeGreaterThan(0);
-
-      // Check for the presence of clickable text elements (terms and learn more links)
-      const clickableElements = screen.getAllByText(
-        /mocked_rewards.onboarding.step4_legal_disclaimer_[24]/,
-      );
-      expect(clickableElements.length).toBeGreaterThan(0);
-    });
-
-    it('should show progress indicator with currentStep set to 4', () => {
-      renderWithProviders(<OnboardingStep4 />);
-      expect(screen.getByTestId('progress-indicator-container')).toBeDefined();
-    });
-  });
-
-  describe('referral code validation', () => {
-    it('should show validation loading state when validating referral code', () => {
-      mockUseValidateReferralCode.isValidating = true;
-      mockUseValidateReferralCode.referralCode = 'TEST123';
-
-      renderWithProviders(<OnboardingStep4 />);
-
-      // Should show loading text
-      expect(
-        screen.getByText(
-          'mocked_rewards.onboarding.step4_title_referral_validating',
-        ),
-      ).toBeDefined();
-    });
-
-    it('should show success icon when referral code is valid', () => {
-      mockUseValidateReferralCode.referralCode = 'VALID123';
-      mockUseValidateReferralCode.isValid = true;
-
-      renderWithProviders(<OnboardingStep4 />);
-
-      // Should show confirmation icon in the text field accessory
-      expect(screen.getByTestId('step-4-image')).toBeDefined();
-    });
-
-    it('should show error icon when referral code is invalid', () => {
-      mockUseValidateReferralCode.referralCode = 'INVALID';
-      mockUseValidateReferralCode.isValid = false;
-
-      renderWithProviders(<OnboardingStep4 />);
-
-      // TextField should be present and show error state
-      expect(screen.getByDisplayValue('INVALID')).toBeDefined();
-    });
-
-    it('should call setReferralCode when input value changes', () => {
-      renderWithProviders(<OnboardingStep4 />);
-
-      const textField = screen.getByDisplayValue('');
-      fireEvent.changeText(textField, 'NEWCODE');
-
-      expect(mockSetReferralCode).toHaveBeenCalledWith('NEWCODE');
-    });
-  });
-
-  describe('opt-in flow', () => {
-    it('should call optin with referral code when next button is pressed', () => {
-      // Set up the mock before rendering
-      Object.assign(mockUseValidateReferralCode, {
-        referralCode: 'TEST123',
-        isValid: true,
-      });
-
-      renderWithProviders(<OnboardingStep4 />);
-
-      const nextButton = screen.getByText(
-        'mocked_rewards.onboarding.step4_confirm',
-      );
-      fireEvent.press(nextButton);
-
-      expect(mockOptin).toHaveBeenCalledWith({
-        referralCode: 'TEST123',
-        isPrefilled: false,
-        bulkLink: false,
-      });
-    });
-
-    it('should show loading state during opt-in', () => {
-      mockUseOptin.optinLoading = true;
-
-      renderWithProviders(<OnboardingStep4 />);
-
-      expect(
-        screen.getByText('mocked_rewards.onboarding.step4_confirm_loading'),
-      ).toBeDefined();
-    });
-
-    it('should disable next button when referral code is invalid', () => {
-      mockUseValidateReferralCode.referralCode = 'INVALID';
-      mockUseValidateReferralCode.isValid = false;
-
-      renderWithProviders(<OnboardingStep4 />);
-
-      const nextButton = screen.getByText(
-        'mocked_rewards.onboarding.step4_confirm',
-      );
-      // Button should be disabled - in real implementation this would be tested via accessibility states
-      expect(nextButton).toBeDefined();
-    });
-  });
-
-  describe('navigation', () => {
-    it('should have swipe gestures disabled', () => {
-      renderWithProviders(<OnboardingStep4 />);
-
-      const container = screen.getByTestId('onboarding-step-container');
-      expect(container).toBeDefined();
-
-      // Test that swiping doesn't trigger navigation
-      fireEvent(container, 'panResponderRelease', null, {
-        dx: -60, // Left swipe
-        dy: 10,
-      });
-
-      // Navigation should not have been called
-      expect(mockNavigate).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('external links', () => {
-    it('should open learn more URL when legal link is pressed', () => {
-      const mockOpenURL = jest.spyOn(Linking, 'openURL');
-
-      renderWithProviders(<OnboardingStep4 />);
-
-      const learnMoreLink = screen.getByText(
-        'mocked_rewards.onboarding.step4_legal_disclaimer_2',
-      );
-      fireEvent.press(learnMoreLink);
-
-      expect(mockOpenURL).toHaveBeenCalledWith(expect.any(String));
-    });
-  });
-
-  describe('state combinations', () => {
-    it('should handle loading and validating states simultaneously', () => {
-      mockUseOptin.optinLoading = true;
-      mockUseValidateReferralCode.isValidating = true;
-
-      renderWithProviders(<OnboardingStep4 />);
-
-      // Should prioritize opt-in loading text over validation loading
-      expect(
-        screen.getByText('mocked_rewards.onboarding.step4_confirm_loading'),
-      ).toBeDefined();
-    });
-
-    it('should disable input when opt-in is loading', () => {
-      mockUseOptin.optinLoading = true;
-
-      renderWithProviders(<OnboardingStep4 />);
-
-      const textField = screen.getByDisplayValue('');
-      // TextField should be disabled when loading - this would be tested via props in real implementation
-      expect(textField).toBeDefined();
-    });
-  });
-});
-
-// Integration tests for step flow
-describe('OnboardingStep Integration', () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-  });
-
-  describe('step progression flow', () => {
-    it('should maintain proper step sequence through navigation', () => {
-      // Test step 1 -> step 2 progression
-      const { rerender } = renderWithProviders(<OnboardingStep1 />);
-
-      const nextButton = screen.getByText(
-        'mocked_rewards.onboarding.step_confirm',
-      );
-      fireEvent.press(nextButton);
-
-      expect(mockNavigate).toHaveBeenCalledWith('RewardsOnboarding2');
-
-      // Simulate navigation to step 2
-      rerender(<OnboardingStep2 />);
-      expect(screen.getByTestId('step-2-image')).toBeDefined();
-    });
-  });
-
-  describe('state persistence', () => {
-    it('should dispatch correct actions for state management', () => {
-      renderWithProviders(<OnboardingStep1 />);
-
-      const nextButton = screen.getByText(
-        'mocked_rewards.onboarding.step_confirm',
-      );
-      fireEvent.press(nextButton);
-
-      // Should dispatch action to set active step
-      expect(mockDispatch).toHaveBeenCalledWith(
-        expect.objectContaining({
-          type: expect.any(String),
-        }),
-      );
-    });
   });
 });
